feat: move out functions to a helper file for attention + add tests for getVariant + useAutoUpdatePosition
