+++ packages/attention/src/component.tsx
@@ -5,16 +5,16 @@ import {
   autoUpdatePosition,
   arrowDirectionClassname,
   useRecompute as recompute,
-  Directions,
 } from '@warp-ds/core/attention'
 import { attention as ccAttention } from '@warp-ds/css/component-classes'
-import { ArrowProps, AttentionProps, AttentionVariants } from './props.js'
+import { ArrowProps, AttentionProps } from './props.js'
 import { i18n } from '@lingui/core'
 import { messages as nbMessages } from './locales/nb/messages.mjs'
 import { messages as enMessages } from './locales/en/messages.mjs'
 import { messages as fiMessages } from './locales/fi/messages.mjs'
 import { activateI18n } from '../../i18n.js'
 import IconClose16 from '@warp-ds/icons/react/close-16'
+import { activeAttentionType, getVariant, pointingAtDirection, useAutoUpdatePosition } from '../../_helpers/attention.js'
 
 const variantClasses = {
   callout: {
@@ -33,87 +33,6 @@ const variantClasses = {
     wrapper: ccAttention.popover,
     arrow: ccAttention.arrowPopover,
   },
-}
-
-const getVariant = (variantProps: AttentionVariants) => {
-  return Object.keys(variantClasses).find((b) => !!variantProps[b]) || ''
-}
-export const pointingAtDirection = (actualDirection: Directions) => {
-  switch (opposites[actualDirection]) {
-    case 'top-start':
-    case 'top':
-    case 'top-end':
-      return i18n._({
-        id: 'attention.aria.pointingUp',
-        message: 'pointing up',
-        comment:
-          'Default screenreader message for top direction in the attention component',
-      })
-    case 'right-start':
-    case 'right':
-    case 'right-end':
-      return i18n._({
-        id: 'attention.aria.pointingRight',
-        message: 'pointing right',
-        comment:
-          'Default screenreader message for right direction in the attention component',
-      })
-    case 'bottom-start':
-    case 'bottom':
-    case 'bottom-end':
-      return i18n._({
-        id: 'attention.aria.pointingDown',
-        message: 'pointing down',
-        comment:
-          'Default screenreader message for bottom direction in the attention component',
-      })
-    case 'left-start':
-    case 'left':
-    case 'left-end':
-      return i18n._({
-        id: 'attention.aria.pointingLeft',
-        message: 'pointing left',
-        comment:
-          'Default screenreader message for left direction in the attention component',
-      })
-    default:
-      return ''
-  }
-}
-
-export const activeAttentionType = (props: AttentionProps) => {
-  switch (true) {
-    case props.tooltip:
-      return i18n._({
-        id: 'attention.aria.tooltip',
-        message: 'tooltip',
-        comment:
-          'Default screenreader message for tooltip in the attention component',
-      })
-    case props.callout:
-      return i18n._({
-        id: 'attention.aria.callout',
-        message: 'callout speech bubble',
-        comment:
-          'Default screenreader message for callout speech bubble in the attention component',
-      })
-    case props.popover:
-      return i18n._({
-        id: 'attention.aria.popover',
-        message: 'popover speech bubble',
-        comment:
-          'Default screenreader message for popover speech bubble in the attention component',
-      })
-    case props.highlight:
-      return i18n._({
-        id: 'attention.aria.highlight',
-        message: 'highlighted speech bubble',
-        comment:
-          'Default screenreader message for highlighted speech bubble in the attention component',
-      })
-    default:
-      return ''
-  }
 }
 
 export function Attention(props: AttentionProps) {
@@ -139,7 +58,7 @@ export function Attention(props: AttentionProps) {
 
   const wrapperClasses = classNames(
     ccAttention.base,
-    variantClasses[getVariant(rest)].wrapper
+    variantClasses[getVariant(rest, variantClasses)].wrapper
   )
 
   const [actualDirection, setActualDirection] = useState(placement)
@@ -195,7 +114,6 @@ export function Attention(props: AttentionProps) {
     },
   }
 
-  // TODO: See if we can move this function to the core repo:
   const defaultAriaLabel = () => {
     return `${activeAttentionType(props)} ${!props.noArrow ? pointingAtDirection(actualDirection) : ''}`
   }
@@ -216,16 +134,8 @@ export function Attention(props: AttentionProps) {
     }
   }, [isShowing, props.callout])
 
-  // @ts-ignore
-  useEffect(() => {
-    if (isShowing && targetEl && attentionEl) {
-      // starts the autoUpdate, making sure the attention elements's position stays anchored to the target element
-      const cleanup = autoUpdatePosition(attentionState)
-
-      return cleanup
-    }
-  }, [targetEl, isShowing, attentionEl])
-
+  useAutoUpdatePosition(targetEl, isShowing, attentionEl, autoUpdatePosition, attentionState)
+  
   return (
     <div
       data-testid='attention-el'
@@ -284,7 +194,7 @@ const Arrow = forwardRef<HTMLDivElement, ArrowProps>(
     const arrowClasses = classNames(
       ccAttention.arrowBase,
       ccAttention[`arrowDirection${arrowDirectionClassname(arrowDirection)}`],
-      variantClasses[getVariant(rest)].arrow
+      variantClasses[getVariant(rest, variantClasses)].arrow
     )
 
     return <div ref={ref} className={arrowClasses} />
